Load Metadata

This function loads the current metadata avaliable from the Semantic Priming GitHub Group.

load_metadata(webaddress = "https://rb.gy/l82i9p")

Arguments

webaddress
The default value for webaddress is the current location of the metadata list.

Value

metadata

The metadata list of avaliable datasets

Examples

#Use the following to load the metadata:
metadata <- load_metadata()
head(metadata)

#>    included
#> 6       yes
#> 7       yes
#> 23      yes
#> 47      yes
#> 53      yes
#> 81      yes
#>                                                                                                link
#> 6  https://raw.githubusercontent.com/SemanticPriming/LAB-data/master/French/AlarioEtAl1999App-A.csv
#> 7  https://raw.githubusercontent.com/SemanticPriming/LAB-data/master/French/AlarioEtAl1999App-B.csv
#> 23          https://raw.githubusercontent.com/SemanticPriming/LAB-data/master/English/Amsel2012.csv
#> 47       https://raw.githubusercontent.com/SemanticPriming/LAB-data/master/Persian/Bakhtiar2013.csv
#> 53     https://raw.githubusercontent.com/SemanticPriming/LAB-data/master/Italian/Barbarotto2005.csv
#> 81     https://raw.githubusercontent.com/SemanticPriming/LAB-data/master/German/Birchenough2017.csv
#>             bibtex                           author year
#> 6       Alario1999                 Alario & Ferrand 1999
#> 7     Alario1999_2                 Alario & Ferrand 1999
#> 23       Amsel2012           Amsel, Urbach, & Kutas 2012
#> 47    Bakhtiar2013     Bakhtiar, Nilipour, & Weekes 2013
#> 53  Barbarotto2005 Barbarotto, Laiacona, & Capitani 2005
#> 81 Birchenough2017  Birchenough, Davies, & Connelly 2016
#>                                                                                                                                                              ref_title
#> 6  A set of 400 pictures standardized for French: Norms for name agreement, image agreement, familiarity, visual complexity, image variability, and age of acquisition
#> 7  A set of 400 pictures standardized for French: Norms for name agreement, image agreement, familiarity, visual complexity, image variability, and age of acquisition
#> 23                                                                                                      Perceptual and motor attribute ratings for 559 object concepts
#> 47                                                                                                                       Predictors of timed picture naming in Persian
#> 53                                                                                 Objective versus estimated age of word acquisition: A study of 202 Italian children
#> 81                                                                                                          Rated age-of acquisition norms for over 3,200 German Words
#>                                            ref_journal ref_volume   ref_page
#> 6  Behavior Research Methods, Instruments, & Computers         31   531--552
#> 7  Behavior Research Methods, Instruments, & Computers         31   531--552
#> 23                           Behavior Research Methods         44 1028--1041
#> 47                           Behavior Research Methods         45   834--841
#> 53                           Behavior Research Methods         37   644--650
#> 81                           Behavior Research Methods         49   484--501
#>                      ref_doi  no1 no2          type1
#> 6         10.3758/BF03200732  400  NA       Pictures
#> 7         10.3758/BF03200732  400  NA       Pictures
#> 23 10.3758/s13428-012-0215-z  560  47          Words
#> 47 10.3758/s13428-012-0298-6  200  NA Color drawings
#> 53        10.3758/BF03192735   80  NA       Pictures
#> 81 10.3758/s13428-016-0718-0 3259  NA          Words
#>                                                        ref1 type2
#> 6                                                                
#> 7                                                                
#> 23               McRae, Cree, Seidenberg, & McNorgan (2005) Words
#> 47 Snodgrass & Vanderwart (1980); Rossion & Pourtois (2004)      
#> 53                                                               
#> 81                                                               
#>                        ref2 notes_stim data_name nonling language notes_lang
#> 6                                                      1   French           
#> 7                                                      1   French           
#> 23 Vinson & Vigiocco (2008)      Nouns                 0  English           
#> 47                                                     1  Persian           
#> 53                                                     1  Italian           
#> 81                                                     0   German           
#>    language_glotto notes_glotto population notes_var accuracy ambiguity aoa
#> 6         stan1290                                          0         0   1
#> 7         stan1290                                          0         0   1
#> 23        stan1293                                          0         0   0
#> 47                                                          0         0   1
#> 53        ital1282                                          0         0   1
#> 81        stan1295                                          0         0   1
#>    arousal assoc category cloze complex concrete confusion context dist
#> 6        0     0        0     0       0        0         0       0    0
#> 7        0     0        0     0       0        0         0       0    0
#> 23       0     0        0     0       0        0         0       0    0
#> 47       0     0        0     0       0        0         0       0    0
#> 53       0     0        0     0       0        0         0       0    0
#> 81       0     0        0     0       0        0         0       0    0
#>    dominate easelearn familiar freq gpc identify identifyld identifyn
#> 6         0         0        1    0   0        0          0         0
#> 7         0         0        1    0   0        0          0         0
#> 23        0         0        1    0   0        0          0         0
#> 47        0         0        0    1   0        0          0         1
#> 53        0         0        0    0   0        0          0         1
#> 81        0         0        0    0   0        0          0         0
#>    imageagree imagevar imagine intense letters meaning modality morph nameagree
#> 6           1        1       0       0       0       0        0     0         1
#> 7           1        1       0       0       0       0        0     0         1
#> 23          0        0       0       0       0       0        0     0         0
#> 47          1        0       1       0       1       0        0     0         1
#> 53          0        0       0       0       0       0        0     0         0
#> 81          0        0       0       0       0       0        0     0         0
#>    orthon phonemes phonon pos prime pronounce recall recognition rime rt
#> 6       0        0      0   0     0         0      0           0    0  0
#> 7       0        0      0   0     0         0      0           0    0  0
#> 23      0        0      0   0     0         0      0           0    0  0
#> 47      0        0      0   0     0         0      0           0    0  1
#> 53      0        0      0   0     0         0      0           0    0  0
#> 81      0        0      0   0     0         0      0           0    0  0
#>    semantic sensory sentcomp similar syllables syntax typical valence
#> 6         0       0        0       0         0      0       0       0
#> 7         0       0        0       0         0      0       0       0
#> 23        0       1        0       0         0      0       0       0
#> 47        0       0        0       0         0      0       0       0
#> 53        0       0        0       0         0      0       0       0
#> 81        0       0        0       0         0      0       0       0
#>    visualcomp wordcomp
#> 6           1        0
#> 7           1        0
#> 23          0        0
#> 47          1        0
#> 53          0        0
#> 81          0        0